requirements.txt
falcon==2.0.*
falcon_marshmallow>=0.3.0
falcon-apispec>=0.4.0
falcon-cors>=1.1.7
falcon-jsonify>=1.2
gunicorn>=20.0.4
python-dotenv>=0.13.0
secure>=0.2.1
stoplight>=1.3.0
